Move index gen to surface gen
