+++ src/main/java/com/google/api/codegen/gapic/MainGapicProviderFactory.java
@@ -38,6 +38,7 @@ import com.google.api.codegen.transformer.go.GoGapicSurfaceTransformer;
 import com.google.api.codegen.transformer.java.JavaGapicSurfaceTestTransformer;
 import com.google.api.codegen.transformer.java.JavaGapicSurfaceTransformer;
 import com.google.api.codegen.transformer.nodejs.NodeJSGapicSurfaceTestTransformer;
+import com.google.api.codegen.transformer.nodejs.NodeJSGapicSurfaceTransformer;
 import com.google.api.codegen.transformer.nodejs.NodeJSPackageMetadataTransformer;
 import com.google.api.codegen.transformer.php.PhpGapicSurfaceTransformer;
 import com.google.api.codegen.util.CommonRenderingUtil;
@@ -176,6 +177,7 @@ public class MainGapicProviderFactory
     } else if (id.equals(NODEJS) || id.equals(NODEJS_DOC)) {
       if (generatorConfig.enableSurfaceGenerator()) {
         GapicCodePathMapper nodeJSPathMapper = new NodeJSCodePathMapper();
+        // TODO: Replace mainProvider with surfaceProvider once NodeJS is MVVM ready
         GapicProvider<? extends Object> mainProvider =
             CommonGapicProvider.<Interface>newBuilder()
                 .setModel(model)
@@ -193,6 +195,13 @@ public class MainGapicProviderFactory
                 .setSnippetSetRunner(new CommonSnippetSetRunner(new CommonRenderingUtil()))
                 .setModelToViewTransformer(new NodeJSPackageMetadataTransformer())
                 .build();
+        GapicProvider<? extends Object> surfaceProvider =
+            ViewModelGapicProvider.newBuilder()
+                .setModel(model)
+                .setApiConfig(apiConfig)
+                .setSnippetSetRunner(new CommonSnippetSetRunner(new CommonRenderingUtil()))
+                .setModelToViewTransformer(new NodeJSGapicSurfaceTransformer())
+                .build();
         GapicProvider<? extends Object> clientConfigProvider =
             CommonGapicProvider.<Interface>newBuilder()
                 .setModel(model)
@@ -206,6 +215,7 @@ public class MainGapicProviderFactory
                 .build();
 
         providers.add(mainProvider);
+        providers.add(surfaceProvider);
         providers.add(metadataProvider);
         providers.add(clientConfigProvider);
